clean version with drink add entry and refactory
